Implement auto-discovery of required hardware resources
PIO-USB uses DMA channel, PIO instructions and State Machines to run its
logic. These hardware resources are limited and usually can be shared
between different modules. The current approach is the user to configure
these resources manually and to resolve any possible conflicts. This
approach may not work in all cases, especially in more complex systems.
This patch implements logic for auto-discovery and allocation of
required hardware resources. The Pico SDK has dedicated APIs for that.
The existing approach for manually configuration still works, the
auto-discovery logic is on top of it. How it can be triggered:
 - Instead of specifying a resource in pio_usb_configuration_t, set
   "-1".
 - Auto-discovery is supported of all resources: pio_tx_num, pio_rx_num,
   tx_ch, sm_tx, sm_rx and sm_eop.
 - Mixed mode is also supported - some of the resources can be set
   manually, others can be auto-discovered.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>